+++ gateway/src/apicast/policy/upstream_mtls/upstream_mtls.lua
@@ -8,11 +8,18 @@ local data_url = require('resty.data_url')
 local C = ffi.C
 local get_request = base.get_request
 local open = io.open
+local pairs = pairs
 
+local X509_STORE = require('resty.openssl.x509.store')
+local X509 = require('resty.openssl.x509')
 
 ffi.cdef([[
   int ngx_http_apicast_ffi_set_proxy_cert_key(
     ngx_http_request_t *r, void *cdata_chain, void *cdata_key);
+  int ngx_http_apicast_ffi_set_proxy_ca_cert(
+    ngx_http_request_t *r, void *cdata_ca);
+  int ngx_http_apicast_ffi_set_ssl_verify(
+    ngx_http_request_t *r, int verify, int verify_deph);
 ]])
 
 
@@ -82,6 +89,26 @@ local function read_certificate_key(value, value_type)
 
 end
 
+local function read_ca_certificates(ca_certificates)
+  local valid = false
+  local store = X509_STORE.new()
+  for _,certificate in pairs(ca_certificates) do
+    local cert, err = X509.parse_pem_cert(certificate)
+    if cert then
+      valid = true
+      store:add_cert(cert)
+    else
+      ngx.log(ngx.INFO, "cannot load certificate, err: ", err)
+    end
+  end
+
+  if valid then
+    return store.store
+  end
+
+  store = nil
+end
+
 function _M.new(config)
   local self = new(config)
   if config == nil then
@@ -94,6 +121,9 @@ function _M.new(config)
   self.cert_key = read_certificate_key(
     config.certificate_key,
     config.certificate_key_type or path_type)
+  self.ca_store = read_ca_certificates(config.ca_certificates or {})
+  self.verify = config.verify
+
   return self
 end
 
@@ -114,9 +144,42 @@ function _M.set_certs(cert, key)
   end
 end
 
+function _M.set_ca_cert(r, store)
+  local val = C.ngx_http_apicast_ffi_set_proxy_ca_cert(r, store)
+  if val == ngx.OK then
+    ngx.log(ngx.WARN, "Cannot set a valid trusted CA store")
+    return
+  end
+end
+
+-- All of this happens on balancer because this is subrequest inside APICAst
+--to @upstream, so the request need to be the one that connects to the
+--upstream0
 function _M:balancer(context)
   if self.cert and self.cert_key then
     self.set_certs(self.cert, self.cert_key)
+  end
+
+  if not self.verify then
+    return
+  end
+
+  local r = get_request()
+  if not r then
+    ngx.log(ngx.WARN, "Invalid request")
+    return
+  end
+
+  if not self.ca_store then
+    ngx.log(ngx.WARN, "Set verify without including CA certificates")
+    return
+  end
+
+  self.set_ca_cert(r, self.ca_store)
+
+  local val = C.ngx_http_apicast_ffi_set_ssl_verify(r, ffi.new("int", 1), ffi.new("int", 1))
+  if val ~= ngx.OK then
+    ngx.log(ngx.WARN, "Cannot verify SSL upstream connection")
   end
 end