+++ spec/policy/upstream_mtls/upstream_mtls_spec.lua
@@ -1,11 +1,21 @@
 local upstream_mtls = require("apicast.policy.upstream_mtls")
 local ssl = require('ngx.ssl')
+local open = io.open
+
+local function read_file(path)
+    local file = open(path, "rb")
+    if not file then return nil end
+    local content = file:read "*a" -- *a or *all reads the whole file
+    file:close()
+    return content
+end
 
 describe('Upstream MTLS policy', function()
 
   local certificate_path = 't/fixtures/CA/root-ca.crt'
   local certificate_key_path = 't/fixtures/CA/root-ca.key'
 
+  local certificate_content = read_file(certificate_path)
   -- Set here the const to not use the pakcage ones, if not test will not fail
   -- if changes0
   local path_type = "path"
@@ -119,5 +129,61 @@ describe('Upstream MTLS policy', function()
 
   end)
 
+  describe("CA_Certificates", function()
+
+
+    local config = {}
+    before_each(function()
+      stub.new(upstream_mtls, 'set_ca_cert', function() return true end)
+      config = {
+        certificate = "XXXX",
+        certificate_type = embedded_type,
+        certificate_key = "XXXX",
+        certificate_key_type = embedded_type,
+      }
+    end)
+
+    it("ca_store is nil if no certificates", function()
+      config.ca_certificates = {}
+      local object = upstream_mtls.new(config)
+      assert.same(object.ca_store, nil)
+    end)
+
+    it("ca_store is nil if no valid certificates", function()
+      config.ca_certificates = {"XXXX"}
+      local object = upstream_mtls.new(config)
+      assert.same(object.ca_store, nil)
+    end)
+
+    it("ca_store is cdata if valid ca certificate", function()
+      config.ca_certificates = { certificate_content}
+      local object = upstream_mtls.new(config)
+      assert.same(type(object.ca_store), "cdata")
+    end)
+
+    it("CA certificate is not used if verify is not enabled", function()
+      config.ca_certificates = { certificate_content}
+      config.verify = false
+
+      local object = upstream_mtls.new(config)
+      assert.same(type(object.ca_store), "cdata")
+
+      spy.on(object, "set_ca_cert")
+      object:balancer({})
+      assert.spy(object.set_ca_cert).was_not.called()
+    end)
+
+    it("CA certificate is used if verify is enabled", function()
+      config.ca_certificates = { certificate_content}
+      config.verify = true
+
+      local object = upstream_mtls.new(config)
+      assert.same(type(object.ca_store), "cdata")
+
+      spy.on(object, "set_ca_cert")
+      object:balancer({})
+      assert.spy(object.set_ca_cert).was.called()
+    end)
+  end)
 end)